Remove current/selected day circle draw
The function 'drawDayCircleIndicator' is responsible to draw a
circle for current/selected day in the calendar. This circle is
too big, so I have reduced its radius to .6 to avoid it overlap
with event dots.
But for now I don't want to draw any circle, so this commit
comments it out from the code.

Signed-off-by: Sergio Durand <sergiodurand@gmail.com>